Using grep directly instead of find | grep.
When searching for workbench logging, the `find | grep` initial solution
to this would fail with an exit code 123, indicating that the search
string `/invest-workbench/` wasn't found in one or more of the files
passed to `xargs` ... which is obviously the case, we didn't start
shipping the workbench until quite recently.

To fix this, better to use a single `grep` call with the `_usage` files
excluded.